requirements.txt
asgiref==3.4.1
attrs==21.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-common-helpers==0.9.2
django-crispy-forms==1.13.0
django-import-export==2.6.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
Jinja2==3.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-jsonrpc-server==0.4.0
python-utils==2.5.6
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==3.10.0.2
ujson==4.2.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
